move getGTEEmbeddding to helperUtils.ts
